feat: refactor SearchSection by modularizing components into SearchBar, CollaborationStatement, UniversityLogos, and CourseDropdown for improved maintainability and design
