+++ src/components/landing/search/CollaborationStatement.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+
+export function CollaborationStatement() {
+  return (
+    <div className="text-center w-[75%] mx-auto mb-8">
+      <h2 className="text-2xl lg:text-3xl font-bold text-gray-800 leading-tight">
+        We collaborate with{' '}
+        <span className="text-[#76B729]">350+</span>{' '}
+        leading universities and companies
+      </h2>
+    </div>
+  );
+}
