+++ src/components/landing/search/SearchInput.tsx
@@ -1,0 +1,25 @@
+'use client';
+
+import React from 'react';
+import { Search } from 'lucide-react';
+import { Input } from '@/components/ui/input';
+
+interface SearchInputProps {
+  searchQuery: string;
+  setSearchQuery: (query: string) => void;
+}
+
+export function SearchInput({ searchQuery, setSearchQuery }: SearchInputProps) {
+  return (
+    <div className="flex-[2] relative">
+      <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 w-4 h-4 pointer-events-none" />
+      <Input
+        type="text"
+        placeholder="What do you want to learn?"
+        value={searchQuery}
+        onChange={(e) => setSearchQuery(e.target.value)}
+        className="pl-10 h-10 border-gray-200 focus:ring-[#76b729] text-sm rounded-md bg-[#F2F2F2]"
+      />
+    </div>
+  );
+}
